application/pyproject.toml
[project]
name = "jsc-voice"
version = "0.1.0"
description = "JSC Voice - A simple API to transcript voice or generate voice from text."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "JScdroidDev", email = "19742131+jscoobyced@users.noreply.github.com"}
]
requires-python = ">=3.12"

dependencies = [
    "asyncio>=4.0.0",
    "einops>=0.8.1",
    "einx>=0.3.0",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "omegaconf>=2.3.0",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
    "soxr>=1.0.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "transformers>=4.57.1",
    "websockets>=15.0.1",
]

[project.urls]
"Homepage" = "https://github.com/jscoobyced/jsc-voice"
"Bug Tracker" = "https://github.com/jscoobyced/jsc-voice/issues"
